requirements.txt
aiohttp==3.8.1
fastapi==0.74.1
fastapi_utils==0.2.1
pandas==1.4.1
protobuf==3.19.4
pydantic==1.9.0
python-dotenv==0.19.2
requests==2.27.1
uvicorn==0.17.5
turfpy
gunicorn
